+++ impl.go
@@ -1,19 +1,21 @@
 package draft
 
 import (
+	"encoding/csv"
+	"io"
 	"strings"
+
+	"github.com/rakyll/statik/fs"
 )
 
-func guessImplByProvider(prov string) func(*Component) {
-	return func(com *Component) {
-		if s := strings.TrimSpace(com.Impl); len(s) > 0 {
-			return
-		}
+func setImpl(com *Component) {
+	if s := strings.TrimSpace(com.Impl); len(s) > 0 {
+		return
+	}
 
-		impl := getCloudImpl(prov, com.Kind)
-		if len(impl) > 0 {
-			com.Impl = impl
-		}
+	impl := getCloudImpl(com.Provider, com.Kind)
+	if len(impl) > 0 {
+		com.Impl = impl
 	}
 }
 
@@ -26,27 +28,12 @@ func getCloudImpl(provider, kind string) string {
 	case "azure":
 		return azureImpl()(kind)
 	default:
-		return ""
+		return defaultImpl()(kind)
 	}
 }
 
 func awsImpl() func(string) string {
-	dict := map[string]string{
-		kindBlockStore:        "Elastic Block\nStore (EBS)",
-		kindCDN:               "Cloudfront",
-		kindDNS:               "Route 53",
-		kindNoSQL:             "DynamoDB",
-		kindFileStore:         "Elastic\nFile System (EFS)",
-		kindFunction:          "Lambda",
-		kindGateway:           "API Gateway",
-		kindLBA:               "Elastic\nLoad Balancer",
-		kindContainersManager: "Elastic Container Service\nfor Kubernetes (EKS)",
-		kindCache:             "ElastiCache",
-		kindPubSub:            "SNS",
-		kindObjectStore:       "Simple Storage\nService (S3)",
-		kindRDB:               "Relational Database\nService (RDS)",
-		kindFirewall:          "AWS WAF",
-	}
+	dict, _ := readCsvFile("/aws.csv")
 
 	return func(key string) string {
 		return dict[key]
@@ -54,22 +41,7 @@ func awsImpl() func(string) string {
 }
 
 func googleImpl() func(string) string {
-	dict := map[string]string{
-		kindBlockStore:        "Persistent Disk",
-		kindCDN:               "Cloud CDN",
-		kindDNS:               "Cloud DNS",
-		kindNoSQL:             "Cloud Datastore",
-		kindFileStore:         "Cloud Filestore",
-		kindFunction:          "Cloud Functions",
-		kindGateway:           "Cloud Endpoints",
-		kindLBA:               "Cloud Load Balancing",
-		kindContainersManager: "Google Kubernetes\nEngine",
-		kindCache:             "Cloud Memorystore",
-		kindPubSub:            "Cloud Pub/Sub",
-		kindObjectStore:       "Cloud Storage",
-		kindRDB:               "Cloud SQL",
-		kindFirewall:          "Google Armor",
-	}
+	dict, _ := readCsvFile("/google.csv")
 
 	return func(key string) string {
 		return dict[key]
@@ -77,24 +49,46 @@ func googleImpl() func(string) string {
 }
 
 func azureImpl() func(string) string {
-	dict := map[string]string{
-		kindBlockStore:        "Disk Storage",
-		kindCDN:               "Azure CDN",
-		kindDNS:               "Azure DNS",
-		kindNoSQL:             "Cosmos DB",
-		kindFileStore:         "Azure File Storage",
-		kindFunction:          "Azure Functions",
-		kindGateway:           "Azure API Management",
-		kindLBA:               "Load Balancer",
-		kindContainersManager: "Azure Kubernetes\nService (AKS)",
-		kindCache:             "Redis Caches",
-		kindPubSub:            "Notification Hubs",
-		kindObjectStore:       "Blob Storage",
-		kindRDB:               "SQL Database",
-		kindFirewall:          "Azure Firewall",
-	}
+	dict, _ := readCsvFile("/azure.csv")
 
 	return func(key string) string {
 		return dict[key]
+	}
+}
+
+func defaultImpl() func(string) string {
+	dict, _ := readCsvFile("/default.csv")
+
+	return func(key string) string {
+		return dict[key]
+	}
+}
+
+func readCsvFile(filePath string) (map[string]string, error) {
+	dict := map[string]string{}
+
+	sfs, err := fs.New()
+	if err != nil {
+		return dict, err
+	}
+
+	f, err := sfs.Open(filePath)
+	if err != nil {
+		return dict, err
+	}
+	defer f.Close()
+
+	csvr := csv.NewReader(f)
+
+	for {
+		row, err := csvr.Read()
+		if err != nil {
+			if err == io.EOF {
+				err = nil
+			}
+			return dict, err
+		}
+
+		dict[row[0]] = row[1]
 	}
 }